requirements.txt
fastapi
uvicorn
pandas
databricks-sql
python-dotenv
requests
